Changes we made at 8th light after the first Backbone JS session.
Template now contains a checkbox that shows if it's been completed yet.
Clicking the checkbox triggers a JS event in the Backbone view for
a Task.  The event updates the task's complete field and calls the
Backbone save method which opens an AJAX connection and saves the task
in Rails.